requirements.txt
apache_beam==2.62.0
protobuf==5.26.1
google-cloud-pubsub